Add a release.nix for Hydra builds.
I'm doing source filtering here, because we have a dangling symlink in
this repository (site) and even if we wouldn't have it, it most likely
will lead to Hakyll trying to read the file anyway. For imperative
builds, it will work, but on NixOS, the source tree is in its own
chroot and FS namespace, so Hakyll will fail while trying to read those
files.

As for the rest, there is a call to "cabal exec runghc" in site.hs,
which obviously will only work if we put cabal into buildInputs... which
IMO is kinda pointless, because we won't need it anyway.

So, I'm patching this out until either everyone uses Nix or there are
other thingies[TM] which make it work on systems without Nix.

Also, there are UTF-8 characters in the HTML files, so I'm also using
glibcLocales in buildInputs and set a german locale for the build. It
doesn't fit my personal taste, because I usually try to replace all
those ugly umlauts with entities, but even here people might disagree so
I'm leaving this as is until my OCD kicks in someday ;-)

Note also, that we're creating two build products, one for the generator
and one for the site itself, so if people don't want to compile their
stuff themselves, they could add the Hydra channel and thus use the
binary cache.

This addresses #7, although I'm not sure to what degree (that's open for
discussion, I think).

Signed-off-by: aszlig <aszlig@redmoonstudios.org>