speed up page load by 1s
the map plugin was loaded in a blocking fashion, so the site couldn’t be
displayed until the map was ready.
This commit delays that until the regular page content is displayed,
with `window.onload`.

Additionally, one font file didn’t exist and caused a time out (was
because of a typo).